requirements.txt
Pillow>=9.0.0
mediapipe>=0.10.0
numpy>=1.21.0
msgpack>=1.0.0
optuna>=3.0.0
